requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.2
fpdf==1.7.2
fpdf2==2.8.3
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
groq==0.25.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.3
hnswlib==0.8.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.5
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
llama-cloud==0.1.26
llama-cloud-services==0.6.31
llama-index==0.12.41
llama-index-agent-openai==0.4.10
llama-index-cli==0.4.3
llama-index-core==0.12.41
llama-index-embeddings-huggingface==0.5.4
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.6
llama-index-llms-openai==0.4.5
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-hnswlib==0.3.0
llama-parse==0.6.31
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.1.0
openai==1.86.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
